fix: 招待画面のデザイン修正
